Refactoring : 입력에서 Player 클래스 분리
- 패키지를 만들어 옮기는 작업과 Client 클래스에서 Player 클래스를
  추출하는 두 가지 작업을 한 커밋에 하였습니다.. ㅜㅜ
- Client 클래스는 InputView로 이름을 변경하여 입력 관련된 작업만 하게
  하였습니다.
- Player 클래스를 만들어 player 목록을 가지고 있게 하였습니다.
- 패키지 이동으로는
  - input 과 output은 view 패키지로 이동하였습니다.
  - 이번 미션의 핵심 비즈니스 로직을 담당하는 클래스는 사다리 게임이기
    때문에 관련 클래스인 LadderGame, Line, Player 클래스를 domain
    패키지로 이동하였습니다.
