--- src/main/java/LadderGame/Main.java
@@ -1,10 +1,0 @@
-package LadderGame;
-
-public class Main {
-
-    public static void main(String[] args) {
-        Client client = new Client();
-
-        client.run();
-    }
-}
