+++ src/main/java/ladder/Main.java
@@ -1,0 +1,12 @@
+package ladder;
+
+import ladder.view.InputView;
+
+public class Main {
+
+    public static void main(String[] args) {
+        InputView inputView = new InputView();
+
+        inputView.run();
+    }
+}
