+++ src/main/java/ladder/domain/Player.java
@@ -1,0 +1,21 @@
+package ladder.domain;
+
+import java.util.List;
+
+public class Player {
+
+    private static List<String> playerList;
+
+    public static List<String> getPlayerList() {
+        return playerList;
+    }
+
+    public Player(List<String> playerList) {
+        this.playerList = playerList;
+    }
+
+    public int getPlayerCount() {
+        return playerList.size();
+    }
+
+}
